use size_t instead of unsigned int for all size-ish types

+++ src/builtins.c
@@ -41,7 +41,7 @@ struct object *builtin_len(struct object_list * args) {
 }
 
 struct object *builtin_puts(struct object_list * args) {
-    for (int i=0; i < args->size; i++) {
+    for (size_t i=0; i < args->size; i++) {
         char str[1024] = {'\0'};
         object_to_str(str, args->values[i]);
         printf("%s", str);
