+++ src/object.h
@@ -34,12 +34,12 @@ struct function {
 
 struct compiled_function {
     struct instruction instructions;
-    unsigned int num_locals;
+    size_t num_locals;
 };
 
 struct object_list {
     struct object *values[OBJECT_LIST_MAX_VALUES];
-    unsigned int size;
+    size_t size;
 
     // for linking in pool
     struct object_list *next;
@@ -78,12 +78,12 @@ struct object *make_string_object(char *str1, char *str2);
 struct object *make_error_object(char *format, ...);
 struct object *make_array_object(struct object_list *elements);
 struct object *make_function_object(struct identifier_list *parameters, struct block_statement *body, struct environment *env);
-struct object *make_compiled_function_object(struct instruction *ins, unsigned int num_locals);
+struct object *make_compiled_function_object(struct instruction *ins, size_t num_locals);
 struct object *copy_object(struct object *obj);
 void free_object(struct object *obj);
 void object_to_str(char *str, struct object *obj);
 
-struct object_list *make_object_list(unsigned int cap);
+struct object_list *make_object_list(size_t cap);
 void free_object_list(struct object_list *list);
 void free_object_list_pool();
 void free_object_pool();
